% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcion.R
\name{Freq_table}
\alias{Freq_table}
\title{Frec_table}
\usage{
Freq_table(variable, min, max, amplitude)
}
\arguments{
\item{variable}{variable continua.}

\item{min}{mínimo valor que presenta la variable.}

\item{max}{máximo valor que presenta la variable.}

\item{amplitude}{amplitud deseada para los intervalos.}
}
\value{
Tabla compuesta de los intervalos, las frecuencias absolutas y las frecuencias relativas.
}
\description{
Esta función permite obtener una tabla de frecuencias absolutas y relativas para una variable continua,
con el fin de acompañar al histograma de la variable.
}
\examples{
data(iris)
Freq_table(iris$Sepal.Length, 4.3, 7.9, 0.6)

}
